+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 - Update device type [#63](https://github.com/PostHog/posthog-ios/pull/63)
+- `$active_feature_flags` event should filter non active flags ([#41](https://github.com/PostHog/posthog-android/pull/41))
 
 ## 2.0.4 - 2023-10-05